fix build

+++ Source/relive_lib/GameObjects/BaseAbe.hpp
@@ -5,7 +5,7 @@
 class BaseAbe : public ::BaseAliveGameObject
 {
 public:
-    explicit BaseAbe(s32 resourceArraySize);
+    explicit BaseAbe(s16 resourceArraySize);
     virtual bool IsEnteringOrExitingDoor() const = 0;
     virtual s32 DoorId() const = 0;
 };
